Deprecate obsolete public definitions from twisted.python.compat

Author: mthuurne
Reviewer: wsanchez
Fixes: ticket:9922

+++ src/twisted/scripts/trial.py
@@ -17,7 +17,6 @@ from twisted.application import app
 from twisted.python import usage, reflect, failure
 from twisted.python.filepath import FilePath
 from twisted.python.reflect import namedModule
-from twisted.python.compat import long
 from twisted import plugin
 from twisted.trial import runner, itrial, reporter
 
@@ -341,7 +340,7 @@ class _BasicOptions:
 
     def opt_random(self, option):
         try:
-            self['random'] = long(option)
+            self['random'] = int(option)
         except ValueError:
             raise usage.UsageError(
                 "Argument to --random must be a positive integer")
@@ -350,7 +349,7 @@ class _BasicOptions:
                 raise usage.UsageError(
                     "Argument to --random must be a positive integer")
             elif self['random'] == 0:
-                self['random'] = long(time.time() * 100)
+                self['random'] = int(time.time() * 100)
 
 
     def opt_without_module(self, option):
